Move to AnnotationHighlightsOverlay
